Add fix for sum() not having type info.

+++ groovy3/src/main/groovy/com/simplicityitself/training/BrokenCode.groovy
@@ -24,7 +24,7 @@ class BrokenCode {
     }
 
     def mean(Collection<Integer> numbers) {
-        return numbers.sum() / numbers.size()
+        return (numbers.sum() as int) / numbers.size()
     }
 
     BigDecimal median(Collection<Integer> numbers) {
